Comments and re-organizing
